Move redis processing to auth package
