+++ user-rest-service/domain/userdomain/user_repository.go
@@ -8,5 +8,4 @@ type Repository interface {
 	CreateSignUpUser(user *SignUpUser) error
 	DeleteSignUpUser(signUpUser *SignUpUser) error
 	FindLoginUserByEmail(email vo.Email) (*LoginUser, error)
-	AddSessionID(sessionID string, userID UserID) error
 }
