+++ user-rest-service/infrastructure/persistence/user.go
@@ -6,24 +6,19 @@ import (
 	"golang.org/x/xerrors"
 
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/apierrors"
-	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/config"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/domain/userdomain"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/domain/vo"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence/datasource"
-	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence/db"
+	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence/rdb"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/interfaces/presenter"
 )
 
 type userRepository struct {
-	*db.RedisHandler
-	*db.MySQLHandler
+	*rdb.MySQLHandler
 }
 
-func NewUserRepository(redisHandler *db.RedisHandler, mysqlHandler *db.MySQLHandler) *userRepository {
-	return &userRepository{
-		redisHandler,
-		mysqlHandler,
-	}
+func NewUserRepository(mysqlHandler *rdb.MySQLHandler) *userRepository {
+	return &userRepository{mysqlHandler}
 }
 
 func (r *userRepository) FindSignUpUserByUserID(userID userdomain.UserID) (*userdomain.SignUpUser, error) {
@@ -210,17 +205,4 @@ func (r *userRepository) FindLoginUserByEmail(email vo.Email) (*userdomain.Login
 	loginUser := userdomain.NewLoginUserFromDataSource(userIDVo, nameVo, emailVo, passwordVo)
 
 	return loginUser, nil
-}
-
-func (r *userRepository) AddSessionID(sessionID string, userID userdomain.UserID) error {
-	conn := r.RedisHandler.Pool.Get()
-	defer conn.Close()
-
-	expirationS := int(config.Env.Cookie.Expiration.Seconds())
-
-	if _, err := conn.Do("SET", sessionID, userID.Value(), "EX", expirationS); err != nil {
-		return apierrors.NewInternalServerError(apierrors.NewErrorString("Internal Server Error"))
-	}
-
-	return nil
 }
