+++ user-rest-service/infrastructure/router/router.go
@@ -14,22 +14,24 @@ import (
 	"github.com/rs/cors"
 
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/config"
+	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/auth"
+	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/auth/imdb"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/externalapi"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/externalapi/client"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence"
-	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence/db"
+	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/persistence/rdb"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/interfaces/handler"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/usecase"
 )
 
 func Run() error {
-	redisHandler, err := db.NewRedisHandler()
+	redisHandler, err := imdb.NewRedisHandler()
 	if err != nil {
 		return err
 	}
 	defer redisHandler.Pool.Close()
 
-	mySQLHandler, err := db.NewMySQLHandler()
+	mySQLHandler, err := rdb.NewMySQLHandler()
 	if err != nil {
 		return err
 	}
@@ -37,9 +39,10 @@ func Run() error {
 
 	accountApiHandler := client.NewAccountApiHandler()
 
-	userRepository := persistence.NewUserRepository(redisHandler, mySQLHandler)
+	userRepository := persistence.NewUserRepository(mySQLHandler)
+	sessionStore := auth.NewSessionStore(redisHandler)
 	accountApi := externalapi.NewAccountApi(accountApiHandler)
-	userUsecase := usecase.NewUserUsecase(userRepository, accountApi)
+	userUsecase := usecase.NewUserUsecase(userRepository, sessionStore, accountApi)
 	userHandler := handler.NewUserHandler(userUsecase)
 
 	router := mux.NewRouter()
